+++ crates/web-client/src/models/secret_key.rs
@@ -1,4 +1,3 @@
-use miden_objects::Word as NativeWord;
 use miden_objects::crypto::dsa::rpo_falcon512::SecretKey as NativeSecretKey;
 use rand::SeedableRng;
 use rand::rngs::StdRng;
@@ -7,6 +6,7 @@ use wasm_bindgen_futures::js_sys::Uint8Array;
 
 use crate::models::public_key::PublicKey;
 use crate::models::signature::Signature;
+use crate::models::signing_inputs::SigningInputs;
 use crate::models::word::Word;
 use crate::utils::{deserialize_from_uint8array, serialize_to_uint8array};
 
@@ -35,11 +35,17 @@ impl SecretKey {
         self.0.public_key().into()
     }
 
-    pub fn sign(&self, message: &Word) -> Result<Signature, JsValue> {
-        let native_message: NativeWord = message.into();
+    // TODO: update to sign instead of sign_with_rng once miden-objects uses miden-crypto 0.16
+    pub fn sign(&self, message: &Word) -> Signature {
+        return self.sign_data(&SigningInputs::new_blind(message));
+    }
+
+    // TODO: update to sign instead of sign_with_rng once miden-objects uses miden-crypto 0.16
+    #[wasm_bindgen(js_name = "signData")]
+    pub fn sign_data(&self, signing_inputs: &SigningInputs) -> Signature {
         let mut rng = StdRng::from_os_rng();
-        let signature = self.0.sign_with_rng(native_message, &mut rng);
-        Ok(signature.into())
+        let native_word = signing_inputs.to_commitment().into();
+        self.0.sign_with_rng(native_word, &mut rng).into()
     }
 
     pub fn serialize(&self) -> Uint8Array {