+++ crates/web-client/test/mocha.global.setup.mjs
@@ -107,6 +107,7 @@ before(async () => {
         Rpo256,
         SecretKey,
         Signature,
+        SigningInputs,
         SlotAndKeys,
         SlotAndKeysArray,
         StorageMap,
@@ -181,6 +182,7 @@ before(async () => {
       window.Rpo256 = Rpo256;
       window.SecretKey = SecretKey;
       window.Signature = Signature;
+      window.SigningInputs = SigningInputs;
       window.SlotAndKeys = SlotAndKeys;
       window.SlotAndKeysArray = SlotAndKeysArray;
       window.StorageMap = StorageMap;